+++ src/euroeval/benchmark_modules/fresh.py
@@ -73,9 +73,10 @@ class FreshEncoderModel(HuggingFaceEncoderModel):
             model_config=model_config, allowed_params=self.allowed_params
         )
 
-        # This is already set when calling `super.__init__`, but we need it to get a
-        # value from `self.model_max_length`, so we set it here as well.
+        # These are already set when calling `super().__init__`, but we need them to get
+        # values from `self.model_max_length`, so we set them here as well.
         self.model_config = model_config
+        self.benchmark_config = benchmark_config
 
         model, tokeniser = load_model_and_tokeniser(
             model_config=model_config,
@@ -127,6 +128,8 @@ class FreshEncoderModel(HuggingFaceEncoderModel):
         Returns:
             The vocabulary size of the model.
         """
+        if self.benchmark_config.vocabulary_size is not None:
+            return self.benchmark_config.vocabulary_size
         match self.model_config.model_id:
             case "fresh-xlm-roberta-base":
                 return 250_002
@@ -145,6 +148,8 @@ class FreshEncoderModel(HuggingFaceEncoderModel):
         Returns:
             The maximum context length of the model.
         """
+        if self.benchmark_config.max_context_length is not None:
+            return self.benchmark_config.max_context_length
         match self.model_config.model_id:
             case "fresh-xlm-roberta-base":
                 return 512